Moved getMatchMediaMock() to another file and shared it with the integration test.
